LoginPage/LoginPage/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
cffi==2.0.0
colorama==0.4.6
cryptography==46.0.5
Django==6.0.3
iniconfig==2.3.0
packaging==26.0
pip==25.3
pluggy==1.6.0
pycparser==3.0
Pygments==2.19.2
pytest==9.0.2
sqlparse==0.5.5
tzdata==2025.3
